+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 6.20.1 - Aug 2, 2024
+
+* Fix permissions issue caused by slightly wrong usage of GCP storage client code
+
+## 6.20.0 - Aug 1, 2024
+
+* Adds support for --archive flag on os-update to archive a full scrape to google cloud storage bucket
+
 ## 6.19.3 - Jul 8, 2024
 
 * Add instructions on DB migrations to docs